refactor: move/rename soil_id related files
