--- terraso_backend/tests/soil_id/test_soil_id.py
@@ -1,21 +1,0 @@
-# Copyright © 2024 Technology Matters
-#
-# This program is free software: you can redistribute it and/or modify
-# it under the terms of the GNU Affero General Public License as published
-# by the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-#
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-# GNU Affero General Public License for more details.
-#
-# You should have received a copy of the GNU Affero General Public License
-# along with this program. If not, see https://www.gnu.org/licenses/.
-
-from soil_id.us_soil import list_soils
-
-
-def test_soil_list():
-    list_soils(-101.9733687, 33.81246789)
-    assert True
